[MOD] Add option to skip installing tools in the CI.

+++ ci/windows-build.ps1
@@ -17,30 +17,32 @@ Remove-Item "external-${env:PLATFORM}-${env:COMPILER}.zip"
 Move-Item external* binaries
 Write-Output "Installed external dependencies"
 
-Write-Output "Installing OpenSSL"
-Invoke-WebRequest "${env:OPENSSL_URL}" -OutFile "OpenSSL.msi"
-Start-Process msiexec.exe -Wait -ArgumentList '/i OpenSSL.msi /l OpenSSL-install.log /qn'
-Remove-Item OpenSSL.msi
-Remove-Item OpenSSL-install.log
-Write-Output "Installed OpenSSL"
+if ("${env:INSTALL_TOOLS}" -eq "YES") {
+    Write-Output "Installing OpenSSL"
+    Invoke-WebRequest "${env:OPENSSL_URL}" -OutFile "OpenSSL.msi"
+    Start-Process msiexec.exe -Wait -ArgumentList '/i OpenSSL.msi /l OpenSSL-install.log /qn'
+    Remove-Item OpenSSL.msi
+    Remove-Item OpenSSL-install.log
+    Write-Output "Installed OpenSSL"
 
-Write-Output "Installing Doxygen"
-New-Item -ItemType directory -Path doxygen | Out-Null
-Set-Location doxygen
-Invoke-WebRequest "${env:DOXYGEN_URL}" -OutFile "doxygen.zip"
-7z x doxygen.zip
-Remove-Item doxygen.zip
-Set-Location "${ROOT_DIR}"
-Write-Output "Installed Doxygen"
+    Write-Output "Installing Doxygen"
+    New-Item -ItemType directory -Path doxygen | Out-Null
+    Set-Location doxygen
+    Invoke-WebRequest "${env:DOXYGEN_URL}" -OutFile "doxygen.zip"
+    7z x doxygen.zip
+    Remove-Item doxygen.zip
+    Set-Location "${ROOT_DIR}"
+    Write-Output "Installed Doxygen"
 
-Write-Output "Installing Qt"
-New-Item -ItemType directory -Path "${env:QT_EXTRACT_DIR}" | Out-Null
-Set-Location "${env:QT_EXTRACT_DIR}"
-Invoke-WebRequest "${env:QT_URL}" -OutFile "Qt.7z"
-7z x "Qt.7z"
-Remove-Item Qt.7z
-Set-Location "${ROOT_DIR}"
-Write-Output "Installed Qt"
+    Write-Output "Installing Qt"
+    New-Item -ItemType directory -Path "${env:QT_EXTRACT_DIR}" | Out-Null
+    Set-Location "${env:QT_EXTRACT_DIR}"
+    Invoke-WebRequest "${env:QT_URL}" -OutFile "Qt.7z"
+    7z x "Qt.7z"
+    Remove-Item Qt.7z
+    Set-Location "${ROOT_DIR}"
+    Write-Output "Installed Qt"
+}
 
 New-Item -ItemType directory -Path build | Out-Null
